Fix the QFL soft label's dtype, and test the loss under autocast
`assign_targets` builds everything from the batch, so its tensors are fp32,
while the model's outputs are fp16 under autocast. The two meet in the QFL soft
label and `scatter_` will not mix them:

  RuntimeError: scatter(): Expected self.dtype to be equal to src.dtype

killed the run at the first step. `torch.where` in the same loop had been
promoting silently all along, which is why nothing upstream complained.

The CPU smoke test runs with autocast off and cannot see any of this, so the
real fix is the test: `test_loss_under_autocast` builds a small model, runs the
whole forward and loss inside `torch.autocast("cpu", bfloat16)`, backprops, and
checks that every parameter received a gradient - which also covers the DDP
"unused parameter" failure mode that cost the first run. Reverting the one-line
fix reproduces the exact RuntimeError, so the test earns its place.

+++ src/train/losses.py
@@ -419,7 +419,11 @@ class SpanLoss:
                 # assigned class and on the agnostic channel. Negatives keep 0,
                 # which is what gives the head something to calibrate against.
                 q = (iou.detach().clamp(0, 1) * sel).unsqueeze(-1)
-                soft = torch.zeros_like(out["cls"][lvl])
+                # `iou` is fp32 (the targets from `assign_targets` are, and
+                # `torch.where` promotes to them) while `out["cls"]` is fp16
+                # under autocast, and `scatter_` will not mix the two. Build the
+                # soft label at the target's dtype; the losses below promote.
+                soft = torch.zeros_like(out["cls"][lvl], dtype=q.dtype)
                 soft.scatter_(2, tgt["cls_idx"][lvl].unsqueeze(-1), q)
                 soft[..., self.n_class] = q.squeeze(-1)
                 l_cls = l_cls + quality_focal(out["cls"][lvl], soft,
